Complete moving lib out
