+++ contracts/mocks/core/lib/SignatureLibraryMock.sol
@@ -1,0 +1,43 @@
+/*
+    Copyright 2018 Set Labs Inc.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+
+pragma solidity 0.4.25;
+
+import { SignatureLibrary } from "../../../core/lib/SignatureLibrary.sol";
+
+/**
+ * @title SignatureLibraryMock
+ * @author Set Protocol
+ *
+ * The Signature Library contains functions for validating signatures
+ *
+ */
+ contract SignatureLibraryMock {
+    function testValidateSignature(
+        bytes32 _orderHash,
+        address _signerAddress,
+        bytes _signature
+    )
+        public
+        pure
+    {
+        return SignatureLibrary.validateSignature(
+            _orderHash,
+            _signerAddress,
+            _signature
+        );
+    }
+ }
