+++ test/contracts/core/extensions/coreIssuanceOrder.spec.ts
@@ -98,8 +98,7 @@ contract('CoreIssuanceOrder', accounts => {
     let subjectRequiredComponents: Address[];
     let subjectRequiredComponentAmounts: BigNumber[];
     let subjectQuantityToFill: BigNumber;
-    let subjectVSignature: BigNumber;
-    let subjectSigBytes: Bytes[];
+    let subjectSignature: Bytes;
     let subjectExchangeOrdersData: Bytes;
 
     let relayerAddress: Address;
@@ -201,7 +200,8 @@ contract('CoreIssuanceOrder', accounts => {
       } as IssuanceOrder;
 
       orderHash = SetUtils.hashOrderHex(issuanceOrder);
-      const signature = await setUtils.signMessage(orderHash, issuanceOrderMaker, false);
+      const ecSignature = await setUtils.signMessage(orderHash, issuanceOrderMaker, false);
+      const signature = setUtils.convertSigToHex(ecSignature);
 
       // Create Taker Wallet transfer for the first component
       takerWalletOrder = {
@@ -266,8 +266,7 @@ contract('CoreIssuanceOrder', accounts => {
       subjectRequiredComponents = issuanceOrder.requiredComponents;
       subjectRequiredComponentAmounts = issuanceOrder.requiredComponentAmounts;
       subjectQuantityToFill = issuanceOrder.quantity;
-      subjectVSignature = new BigNumber(signature.v);
-      subjectSigBytes = [signature.r, signature.s];
+      subjectSignature = signature;
       subjectExchangeOrdersData = setUtils.generateSerializedOrders([zeroExOrder, takerWalletOrder, kyberTrade]);
       subjectCaller = issuanceOrderTaker;
     });
@@ -279,8 +278,7 @@ contract('CoreIssuanceOrder', accounts => {
         subjectRequiredComponents,
         subjectRequiredComponentAmounts,
         subjectQuantityToFill,
-        subjectVSignature,
-        subjectSigBytes,
+        subjectSignature,
         subjectExchangeOrdersData,
         { from: subjectCaller, gas: DEFAULT_GAS },
       );
